src/python/requirements.txt
pandas==2.2.3
psycopg2_binary==2.9.10
Requests==2.32.3
SQLAlchemy==2.0.36
